@@ -6,19 +6,22 @@ import org.bukkit.block.Biome;
 import org.bukkit.entity.Player;
 
 import java.util.Collections;
-import java.util.EnumSet;
+import java.util.HashSet; // Added
+import java.util.List;
 import java.util.Set;
+// import java.util.EnumSet; // Removed
 
 public class BiomeCondition implements SpawnCondition {
 
     private final Set<Biome> allowedBiomes;
 
     public BiomeCondition(Set<Biome> allowedBiomes) {
-        this.allowedBiomes = Collections.unmodifiableSet(EnumSet.copyOf(allowedBiomes));
+        this.allowedBiomes = Collections.unmodifiableSet(new HashSet<>(allowedBiomes)); // Changed to HashSet
     }
 
     public BiomeCondition(Biome... biomes) {
-        this.allowedBiomes = Collections.unmodifiableSet(EnumSet.copyOf(Set.of(biomes)));
+        // Explicitly create a List, which is a Collection, for EnumSet.copyOf
+        this.allowedBiomes = Collections.unmodifiableSet(new HashSet<>(List.of(biomes))); // Changed to HashSet
     }
 
     @Override
