@@ -1,6 +1,6 @@
 package com.x1f4r.mmocraft.world.resourcegathering.service;
 
-import com.x1f4r.mmocraft.MMOCraftPlugin;
+import com.x1f4r.mmocraft.core.MMOCraftPlugin;
 import com.x1f4r.mmocraft.item.service.CustomItemRegistry; // Not directly used in this impl, but good for context
 import com.x1f4r.mmocraft.loot.service.LootService;       // Not directly used in this impl, but good for context
 import com.x1f4r.mmocraft.util.LoggingUtil;
@@ -12,6 +12,7 @@ import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.inventory.ItemStack; // If we drop items directly, not via LootService here
 
+import java.util.Collections;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
